V0.3

+++ MyTelegramBot/BotLogic/TelegramFileDownloader.cs
@@ -8,6 +8,9 @@ namespace MyTelegramBot
     {
         private readonly IMessageSender _messageSender;
         private readonly IFilePathProvider _filePathProvider;
+        public event Action? DownloadStarted;
+        public event Action? DownloadCompleted;
+
         private ITelegramBotClient BotClient { get; set; }
         private Update Update { get; set; }
 
@@ -22,6 +25,8 @@ namespace MyTelegramBot
         {
             if (Update.Message.Document != null)
             {
+                OnDownloadStarted();
+
                 try
                 {
                     var fileId = Update.Message.Document.FileId;
@@ -31,27 +36,44 @@ namespace MyTelegramBot
 
                     string destinationFilePath = _filePathProvider.GetDestinationFilePath(fileName);
 
+                    if (System.IO.File.Exists(destinationFilePath))
+                    {
+                        await _messageSender.SendTextMessageAsync(BotClient, Update.Message.Chat.Id, "File already exists. Skipping download.");
+                        return;
+                    }
+
                     await using Stream fileStream = System.IO.File.Create(destinationFilePath);
 
                     if (!string.IsNullOrEmpty(filePath))
                     {
                         await BotClient.DownloadFileAsync(filePath, fileStream);
-                        await _messageSender.SendTextMessageAsync(BotClient, Update.Message.Chat.Id, "Document downloaded!");
+                        OnDownloadCompleted();
                     }
+
                     else
                     {
                         await _messageSender.SendTextMessageAsync(BotClient, Update.Message.Chat.Id, "Ivalid File path");
                     }
+
                 }
                 catch (Exception ex)
                 {
                     await _messageSender.SendTextMessageAsync(BotClient, Update.Message.Chat.Id, $"An error occurred: {ex.Message}");
                 }
+
             }
             else
             {
                 await _messageSender.SendTextMessageAsync(BotClient, Update.Message.Chat.Id, "Document was null or not provided.");
             }
+        }
+        protected virtual void OnDownloadStarted()
+        {
+            DownloadStarted?.Invoke();
+        }
+        protected virtual void OnDownloadCompleted()
+        {
+            DownloadCompleted?.Invoke();
         }
     }
 }
